Also use the user search base when authenticating
When AM is used in APIM for user authentication with an LDAP provider,
the authentication fails. The reason is that the user search base
setting from the LDAP provider configuration is ignored and thus the
user that is being authenticated is searched at base DN level, instead
of at the "base DN + user search base" level, as it should be.

The user search base is properly handled in the userSearchExecutor and
groupSearchExecutor methods/beans, but the respective code is missing
from authenticator() in the LdapAuthenticationProviderConfiguration
class. This patch corrects the problem.